+++ src/main/kotlin/br/com/storehouse/data/model/ProdutoDto.kt
@@ -1,5 +1,6 @@
 package br.com.storehouse.data.model
 
+import java.math.BigDecimal
 import java.util.*
 
 // ProdutoDto.kt
@@ -7,8 +8,8 @@ data class ProdutoDto(
     val codigoBarras: String,
     val tipoId: UUID,
     val nome: String,
-    val preco: Double,
-    val precoCusto: Double,
+    val preco: BigDecimal,
+    val precoCusto: BigDecimal,
     val estoque: Int,
     val descricaoCampos: Map<String, Any>? = null, // ← Novo campo
     val imagemUrl: String? = null // Referencia para imagem remota
